Rename package from part to provider
