+++ src/test/java/uk/gov/hmcts/divorce/document/content/ApplicationJointTemplateContentTest.java
@@ -8,8 +8,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
 import uk.gov.hmcts.divorce.divorcecase.model.Application;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
-import uk.gov.hmcts.divorce.document.content.part.ApplicantTemplateDataProvider;
-import uk.gov.hmcts.divorce.document.content.part.ApplicationTemplateDataProvider;
+import uk.gov.hmcts.divorce.document.content.provider.ApplicantTemplateDataProvider;
+import uk.gov.hmcts.divorce.document.content.provider.ApplicationTemplateDataProvider;
 
 import java.time.LocalDate;
 import java.util.Map;
